# lib/restapi/openapi.yaml
openapi: 3.1.0
info:
  title: transliter REST API
  version: 1.0.0
  summary: Asynchronous translation jobs
  description: |
    Queues translation jobs, runs them through a user-managed OpenAI-compatible
    model server, and retains owner-scoped job history. Inbound API keys are
    optional: when TRANSLITER_SERVER_API_KEYS is unset, job endpoints are open
    and jobs are owned by anonymous.

    Queue delivery is at-least-once. A model call can be repeated if a worker
    stops while a running job is being processed.

    translation.glossary is required on every job and catalog preview request.
    An empty object means no term constraints. TranslateGemma's official chat
    template rejects non-empty glossaries; Hy-MT2 accepts them on any prompt kind.
  license:
    name: MIT
    identifier: MIT
externalDocs:
  description: Project documentation
  url: https://github.com/snowmerak/transliter
servers:
  - url: /
    description: Current server
tags:
  - name: Jobs
    description: Asynchronous translation jobs (API key optional)
  - name: Models
    description: Upstream OpenAI-compatible provider model list proxy
  - name: Model catalogs
    description: Built-in prompt adapters, capabilities, and dry-run preview
  - name: System
    description: Process health and API documentation
paths:
  /v1/jobs:
    post:
      tags: [Jobs]
      operationId: createTranslationJob
      summary: Create a translation job
      description: Validates, stores, and enqueues a translation job.
      security:
        - {}
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
            examples:
              text:
                summary: English to Korean with Hy-MT2
                value:
                  model_catalog: hymt2-30b-a3b
                  model: hy-mt2-local
                  profile: deterministic
                  translation:
                    source: The service is ready.
                    source_language: English
                    target_language: Korean
                    kind: text
                    glossary: {}
      responses:
        "202":
          description: Job stored and enqueued
          headers:
            Location:
              description: Relative URL for the created job
              schema:
                type: string
                example: /v1/jobs/52d14df7b8b5a7d231b2295790994332
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    get:
      tags: [Jobs]
      operationId: listTranslationJobs
      summary: List previous jobs for the caller
      security:
        - {}
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: limit
          in: query
          description: Maximum jobs to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: before
          in: query
          description: Return jobs created before this RFC3339 timestamp
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Owner-scoped job history in descending creation order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobHistory"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/jobs/{id}:
    get:
      tags: [Jobs]
      operationId: getTranslationJob
      summary: Get one translation job
      description: Cross-owner reads return 404 without revealing job existence.
      security:
        - {}
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Cryptographically random 128-bit job ID
          schema:
            type: string
            pattern: "^[0-9a-f]{32}$"
      responses:
        "200":
          description: Current job state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/model-catalogs:
    get:
      tags: [Model catalogs]
      operationId: listModelCatalogs
      summary: List built-in model catalogs
      description: Returns discovery metadata for every compiled prompt adapter.
      security: []
      responses:
        "200":
          description: Model catalog list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelCatalogList"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/model-catalogs/{id}:
    get:
      tags: [Model catalogs]
      operationId: getModelCatalog
      summary: Get one model catalog
      description: Returns capabilities, languages, profiles, and generation options.
      security: []
      parameters:
        - $ref: "#/components/parameters/ModelCatalogIDParam"
      responses:
        "200":
          description: Model catalog detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelCatalogDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/model-catalogs/{id}/preview:
    post:
      tags: [Model catalogs]
      operationId: previewModelCatalog
      summary: Preview BuildInput and Options
      description: |
        Renders the model input and generation options for a translation request
        without calling an inference server.
      security: []
      parameters:
        - $ref: "#/components/parameters/ModelCatalogIDParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelCatalogPreviewRequest"
      responses:
        "200":
          description: Rendered model input and options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelCatalogPreviewResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/models:
    get:
      tags: [Models]
      operationId: listProviderModels
      summary: List upstream provider models
      description: |
        Proxies GET {provider}/v1/models from the configured OpenAI-compatible
        inference server and returns the upstream status and body unchanged.
      security: []
      responses:
        "200":
          description: Upstream model list
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "502":
          description: Provider models endpoint unreachable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /healthz:
    get:
      tags: [System]
      operationId: getHealth
      summary: Get process health
      security: []
      responses:
        "200":
          description: Process is running
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok
  /openapi.json:
    get:
      tags: [System]
      operationId: getOpenAPIJSON
      summary: Download the OpenAPI document as JSON
      security: []
      responses:
        "200":
          description: OpenAPI 3.1 document
          content:
            application/json:
              schema:
                type: object
  /openapi.yaml:
    get:
      tags: [System]
      operationId: getOpenAPIYAML
      summary: Download the OpenAPI document as YAML
      security: []
      responses:
        "200":
          description: OpenAPI 3.1 document
          content:
            application/yaml:
              schema:
                type: string
  /docs:
    get:
      tags: [System]
      operationId: getScalarDocumentation
      summary: Open the interactive Scalar API reference
      security: []
      responses:
        "200":
          description: Scalar HTML application
          content:
            text/html:
              schema:
                type: string
  /ui:
    get:
      tags: [System]
      operationId: getWebUIRedirect
      summary: Redirect to the job console UI
      security: []
      responses:
        "301":
          description: Redirects to /ui/
  /ui/:
    get:
      tags: [System]
      operationId: getWebUI
      summary: Open the simple job console web UI
      description: |
        Public HTML console for composing translation jobs. API key input is
        optional and only sent when filled. Glossary is required; use {}.
      security: []
      responses:
        "200":
          description: Job console HTML application
          content:
            text/html:
              schema:
                type: string
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Optional client API key when TRANSLITER_SERVER_API_KEYS is configured
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Optional client API key when TRANSLITER_SERVER_API_KEYS is configured
  parameters:
    ModelCatalogIDParam:
      name: id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ModelCatalogID"
  responses:
    BadRequest:
      description: Invalid JSON, query parameter, or request shape
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid client API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Job not found for the authenticated owner
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnprocessableEntity:
      description: Unsupported model, language, profile, or translation contract
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: Queue or store is unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    CreateJobRequest:
      type: object
      additionalProperties: false
      required: [model_catalog, translation]
      properties:
        model_catalog:
          $ref: "#/components/schemas/ModelCatalogID"
        model:
          type: string
          description: Model name or alias understood by the inference server
        profile:
          $ref: "#/components/schemas/OptionProfile"
        translation:
          $ref: "#/components/schemas/TranslationRequest"
    TranslationRequest:
      type: object
      additionalProperties: false
      required: [target_language, glossary]
      properties:
        source:
          type: string
        source_language:
          $ref: "#/components/schemas/Language"
        target_language:
          $ref: "#/components/schemas/Language"
        kind:
          $ref: "#/components/schemas/PromptKind"
        glossary:
          type: object
          description: |
            Required term map from source wording to required target wording.
            Use an empty object when no glossary terms apply. TranslateGemma
            official templates reject non-empty glossaries.
          additionalProperties:
            type: string
        style:
          type: string
        audience:
          type: string
        translatable_attributes:
          type: array
          items:
            type: string
        delimiters:
          type: array
          items:
            type: string
    ModelCatalogID:
      type: string
      description: Built-in prompt/options adapter id
      enum:
        - hymt2-1.8b
        - hymt2-7b
        - hymt2-30b-a3b
        - translategemma-4b
        - translategemma-12b
        - translategemma-27b
    OptionProfile:
      type: string
      default: official
      enum: [official, deterministic]
    PromptKind:
      type: string
      default: text
      enum:
        - text
        - markdown
        - json
        - yaml
        - html_xml
        - mixed_code
        - style_audience
        - segmented
    Language:
      type: string
      description: Typed language name supported by the selected model
      examples: [English, Korean, Japanese]
    Job:
      type: object
      required:
        - id
        - status
        - request
        - created_at
        - updated_at
        - expires_at
      properties:
        id:
          type: string
          pattern: "^[0-9a-f]{32}$"
        status:
          $ref: "#/components/schemas/JobStatus"
        request:
          $ref: "#/components/schemas/CreateJobRequest"
        result:
          $ref: "#/components/schemas/JobResult"
        error:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
    JobStatus:
      type: string
      enum: [queued, running, succeeded, failed]
    JobResult:
      type: object
      required: [translation]
      properties:
        translation:
          type: string
        model:
          type: string
          description: Inference-server model name that produced the result
        finish_reason:
          type: string
        prompt_tokens:
          type: integer
          minimum: 0
        output_tokens:
          type: integer
          minimum: 0
        total_tokens:
          type: integer
          minimum: 0
    JobHistory:
      type: object
      required: [jobs]
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
    ModelCatalogList:
      type: object
      required: [model_catalogs]
      properties:
        model_catalogs:
          type: array
          items:
            $ref: "#/components/schemas/ModelCatalogSummary"
    ModelCatalogSummary:
      type: object
      required: [id, family, capabilities, profiles, languages]
      properties:
        id:
          $ref: "#/components/schemas/ModelCatalogID"
        family:
          type: string
        parameters:
          type: string
        repository:
          type: string
        gguf_repository:
          type: string
        capabilities:
          $ref: "#/components/schemas/ModelCatalogCapabilities"
        profiles:
          type: array
          items:
            $ref: "#/components/schemas/OptionProfile"
        languages:
          type: array
          items:
            $ref: "#/components/schemas/Language"
    ModelCatalogDetail:
      allOf:
        - $ref: "#/components/schemas/ModelCatalogSummary"
        - type: object
          required: [profile_options]
          properties:
            profile_options:
              type: object
              additionalProperties:
                $ref: "#/components/schemas/GenerationOptions"
    ModelCatalogCapabilities:
      type: object
      required:
        - prompt_kinds
        - requires_source_language
        - structured_user_content
        - auxiliary_fields
      properties:
        prompt_kinds:
          type: array
          items:
            $ref: "#/components/schemas/PromptKind"
        requires_source_language:
          type: boolean
        structured_user_content:
          type: boolean
        max_input_tokens:
          type: integer
          minimum: 0
        auxiliary_fields:
          $ref: "#/components/schemas/ModelCatalogAuxiliaryFields"
    ModelCatalogAuxiliaryFields:
      type: object
      required: [glossary, style, audience, translatable_attributes, delimiters]
      properties:
        glossary:
          type: boolean
        style:
          type: boolean
        audience:
          type: boolean
        translatable_attributes:
          type: boolean
        delimiters:
          type: boolean
    GenerationOptions:
      type: object
      properties:
        temperature:
          type: number
        top_p:
          type: number
        top_k:
          type: integer
        repetition_penalty:
          type: number
        do_sample:
          type: boolean
        max_output_tokens:
          type: integer
        stop:
          type: array
          items:
            type: string
        provenance:
          type: string
    ModelCatalogPreviewRequest:
      type: object
      additionalProperties: false
      required: [translation]
      properties:
        profile:
          $ref: "#/components/schemas/OptionProfile"
        translation:
          $ref: "#/components/schemas/TranslationRequest"
    ModelCatalogPreviewResponse:
      type: object
      required: [model_catalog, profile, options, input]
      properties:
        model_catalog:
          $ref: "#/components/schemas/ModelCatalogID"
        profile:
          $ref: "#/components/schemas/OptionProfile"
        options:
          $ref: "#/components/schemas/GenerationOptions"
        input:
          $ref: "#/components/schemas/ModelInput"
    ModelInput:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ModelMessage"
    ModelMessage:
      type: object
      required: [role]
      properties:
        role:
          type: string
          const: user
        content:
          description: Plain string for Hy-MT2, or structured parts for TranslateGemma
          oneOf:
            - type: string
            - type: array
              items:
                type: object
                required: [type, text]
                properties:
                  type:
                    type: string
                  text:
                    type: string
                  source_lang_code:
                    type: string
                  target_lang_code:
                    type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [status, message]
          properties:
            status:
              type: integer
            message:
              type: string